+++ app/src/src_nonfreenet/org/breezyweather/sources/openweather/OpenWeatherApi.kt
@@ -21,7 +21,6 @@ import org.breezyweather.sources.openweather.json.OpenWeatherAirPollutionResult
 import org.breezyweather.sources.openweather.json.OpenWeatherForecast
 import org.breezyweather.sources.openweather.json.OpenWeatherForecastResult
 import retrofit2.http.GET
-import retrofit2.http.Path
 import retrofit2.http.Query
 
 /**